pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-video"
version = "0.1.0"
description = "Pure-local ComfyUI long-video orchestration CLI"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "pydantic>=2.7",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
mcp = [
  "mcp>=1.20",
  "pydantic-settings>=2.0",
  "openai-whisper>=20231117",
  "Pillow>=10.0",
]

[project.scripts]
ai-video = "ai_video.cli:main"
video-analysis-mcp = "ai_video_mcp.server:mcp"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
